make batchpipe a package
