+++ pkg/util/metric/testutil_test.go
@@ -50,31 +50,6 @@ func waitWgTimeout(wg *sync.WaitGroup, after time.Duration) error {
 	}
 }
 
-// waitChTimeout returns an error if:
-// 1. OnRecvCheck returns an error
-// 2. timeout happens before the channel running out or OnRecvCheck asking to stop
-func waitChTimeout[T any](
-	ch <-chan T,
-	onRecvCheck func(element T, closed bool) (goOn bool, err error),
-	after time.Duration,
-) error {
-	timeout := time.After(after)
-	for {
-		select {
-		case <-timeout:
-			return errors.New("timeout")
-		case item, ok := <-ch:
-			goOn, err := onRecvCheck(item, !ok)
-			if err != nil {
-				return err
-			}
-			if !ok || !goOn {
-				return nil
-			}
-		}
-	}
-}
-
 func makeDummyClock(startOffset int64) func() int64 {
 	var tick int64 = startOffset - 1
 	return func() int64 {
